Some more renames to relationship
