+++ datahub-web-react/src/app/entity/ermodelrelationships/ERModelRelationshipEntity.tsx
@@ -6,7 +6,7 @@ import { getDataForEntityType } from '../shared/containers/profile/utils';
 import { GenericEntityProperties } from '../shared/types';
 import { ERModelRelationshipPreviewCard } from './preview/ERModelRelationshipPreviewCard';
 import ermodelrelationshipIcon from '../../../images/ermodelrelationshipIcon.svg';
-import { ERModelRelationshipTab } from '../shared/tabs/ERModelRelation/ERModelRelationTab';
+import { ERModelRelationshipTab } from '../shared/tabs/ERModelRelationship/ERModelRelationshipTab';
 import { useGetErModelRelationshipQuery,  useUpdateErModelRelationshipMutation } from '../../../graphql/ermodelrelationship.generated';
 import { DocumentationTab } from '../shared/tabs/Documentation/DocumentationTab';
 import { PropertiesTab } from '../shared/tabs/Properties/PropertiesTab';
