+++ datahub-web-react/src/app/entity/shared/tabs/ERModelRelationship/ERModelRelationshipTab.less
